+++ opentelemetry-sdk/src/trace/mod.rs
@@ -19,7 +19,13 @@ mod tracer;
 
 pub use config::{config, Config};
 pub use events::SpanEvents;
-pub use id_generator::{aws::XrayIdGenerator, IdGenerator, RandomIdGenerator};
+
+#[deprecated(
+    since = "0.21.3",
+    note = "XrayId Generator has been migrated to the opentelemetry-aws crate"
+)]
+pub use id_generator::aws::XrayIdGenerator;
+pub use id_generator::{IdGenerator, RandomIdGenerator};
 pub use links::SpanLinks;
 pub use provider::{Builder, TracerProvider};
 pub use sampler::{Sampler, ShouldSample};