+++ opentelemetry-sdk/CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
+
 ## vNext
 
 ## v0.22.0
+
+### Deprecated
+
+- XrayIdGenerator in the opentelemetry-sdk has been deprecated and moved to version 0.10.0 of the opentelemetry-aws crate.
 
 ### Added
 
@@ -14,7 +19,6 @@
 Performance Improvement : Creating Spans and LogRecords are now faster, by avoiding expensive cloning of `Resource` for every Span/LogRecord.
 
 ### Changed
-
 - **Breaking**
 [#1313](https://github.com/open-telemetry/opentelemetry-rust/pull/1313)
 [#1350](https://github.com/open-telemetry/opentelemetry-rust/pull/1350)